Add base home page and character page
This commit adds the basic home page with some boilerplate template of what is intended

In addition it include the characters page and an example of linking to the characters page and back to home page etc.

The character page has some non functioning links to pages that have yet to be created. This is known and intentional for now.